requirements.txt
captcha==0.3
certifi==2019.6.16
chardet==3.0.4
dataclasses==0.6
Django==2.2.2
djangorestframework==3.9.4
djongo==1.2.33
idna==2.8
Pillow==6.1.0
pymongo==3.8.0
PyQRCode==1.2.1
pytz==2019.1
requests==2.22.0
sqlparse==0.2.4
urllib3==1.25.3
